Refactor import statements for clarity and robustness
This commit refactors the import statements within the `mbi` package and its tests according to the following principles:

1.  **Internal Relative Imports:** Imports between modules within the `src/mbi/` directory now use explicit relative imports (e.g., `from .domain import Domain`). This clarifies internal dependencies and reduces reliance on `__init__.py` structure.

2.  **Explicit Public API:** `src/mbi/__init__.py` now includes an `__all__` list, explicitly defining the public interface of the package. Imports within `__init__.py` remain relative.

3.  **Explicit Absolute Test Imports:** Imports within the `test/` directory now use explicit absolute paths from the `mbi` root (e.g., `from mbi.domain import Domain`). This makes test dependencies clear and robust.

These changes improve code maintainability and reduce the potential for circular import errors.

Note: Please run the full test suite to ensure no regressions were introduced by this refactoring.
