+++ src/mbi/__init__.py
@@ -7,3 +7,15 @@ from . import estimation
 from . import callbacks
 from . import junction_tree
 from . import marginal_oracles
+
+__all__ = [
+    'Domain',
+    'Dataset',
+    'Factor',
+    'CliqueVector',
+    'LinearMeasurement',
+    'estimation',
+    'callbacks',
+    'junction_tree',
+    'marginal_oracles',
+]
